+++ src/webapp/src/views/Connections.vue
@@ -59,7 +59,6 @@ export default {
   <router-view-layout>
     <div class="container view-body is-widescreen">
       <h2 id="data" class="title">Connections</h2>
-      <p class="subtitle">Integrations and custom data connections</p>
 
       <div class="columns">
         <div class="column">
@@ -68,44 +67,7 @@ export default {
               v-if="getIsLoadingPluginsOfType('extractors')"
               class="progress is-small is-info"
             ></progress>
-            <template v-else>
-              <ExtractorList />
-              <hr />
-              <div class="columns is-vcentered">
-                <div class="column">
-                  <article class="media">
-                    <figure class="media-left">
-                      <p class="image level-item container">
-                        <span class="icon is-large fa-2x has-text-grey-light">
-                          <font-awesome-icon icon="plus"></font-awesome-icon>
-                        </span>
-                      </p>
-                    </figure>
-                    <div class="media-content">
-                      <div class="content">
-                        <p>
-                          <span class="has-text-weight-bold">Custom</span>
-                          <br />
-                          <small>Connect a data source not listed above</small>
-                        </p>
-                      </div>
-                    </div>
-                  </article>
-                </div>
-                <div class="column">
-                  <div class="field is-grouped is-grouped-right">
-                    <a
-                      href="https://www.meltano.com/tutorials/create-a-custom-extractor.html"
-                      target="_blank"
-                      class="button is-text tooltip is-tooltip-left"
-                      data-tooltip="Create your own data source"
-                    >
-                      <span>Learn More</span>
-                    </a>
-                  </div>
-                </div>
-              </div>
-            </template>
+            <ExtractorList v-else />
           </div>
         </div>
       </div>