+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 - [#1852](https://gitlab.com/meltano/meltano/-/issues/1852) Move Pipelines after Connections in navbar
 - [#1850](https://gitlab.com/meltano/meltano/-/issues/1850) Rename Connections tab "Connection" and "Pipeline" buttons to "Edit Connection", and "View Pipeline"
+- [#1856](https://gitlab.com/meltano/meltano/-/issues/1856) Remove "Custom" data source option from UI
 
 ### Fixes